+++ src/Renderer/MeshData.hpp
@@ -8,19 +8,19 @@ struct MeshData
 {
     std::vector<math::Vec3f> vertices;
 	std::vector<math::Vec2f> texCoord;
-	std::vector<unsigned short> elements;
+	std::vector<unsigned int> elements;
     
-    unsigned short addVertex(math::Vec3f pos, 
+    unsigned int addVertex(math::Vec3f pos, 
                              math::Vec2f ptexCoord);
 	
-	void addFace(unsigned short p1, 
-                 unsigned short p2, 
-                 unsigned short p3);
+	void addFace(unsigned int p1, 
+                 unsigned int p2, 
+                 unsigned int p3);
     
-	void addFace(unsigned short p1,
-                 unsigned short p2, 
-                 unsigned short p3, 
-                 unsigned short p4);
+	void addFace(unsigned int p1,
+                 unsigned int p2, 
+                 unsigned int p3, 
+                 unsigned int p4);
     
     void makeCube();
 };
